:recycle: decouple forwarder models from behavior
