+++ internal/engines/forwarders/datadog_test.go
@@ -1,21 +1,17 @@
-package models_test
+package forwarders_test
 
 import (
-	"net/http/httptest"
 	"testing"
 
 	"net/http"
+	"net/http/httptest"
 
+	"link-society.com/flowg/internal/engines/forwarders"
 	"link-society.com/flowg/internal/models"
 )
 
-func TestForwarderHttp_Call(t *testing.T) {
+func TestForwarderDatadog_Call(t *testing.T) {
 	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		fooHeader := r.Header.Get("Foo")
-		if fooHeader != "Bar" {
-			t.Fatalf("unexpected header value: %s", fooHeader)
-		}
-
 		w.WriteHeader(http.StatusOK)
 	}))
 	defer testServer.Close()
@@ -23,30 +19,33 @@ func TestForwarderHttp_Call(t *testing.T) {
 	forwarder := &models.ForwarderV2{
 		Version: 2,
 		Config: models.ForwarderConfigV2{
-			Http: &models.ForwarderHttpV2{
-				Url: testServer.URL,
-				Headers: map[string]string{
-					"Foo": "Bar",
-				},
+			Datadog: &models.ForwarderDatadogV2{
+				Url:    testServer.URL,
+				ApiKey: "apiKey",
 			},
 		},
 	}
 
-	if err := forwarder.Init(t.Context()); err != nil {
+	runtime, err := forwarders.NewRuntime(forwarder)
+	if err != nil {
+		t.Fatalf("failed to create runtime: %v", err)
+	}
+
+	if err := runtime.Init(t.Context()); err != nil {
 		t.Fatalf("failed to initialize forwarder: %v", err)
 	}
 
 	record := models.NewLogRecord(map[string]string{})
-	if err := forwarder.Call(t.Context(), record); err != nil {
+	if err := runtime.Call(t.Context(), record); err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
 
-	if err := forwarder.Close(t.Context()); err != nil {
+	if err := runtime.Close(t.Context()); err != nil {
 		t.Fatalf("failed to close forwarder: %v", err)
 	}
 }
 
-func TestForwarderHttp_Call_Failure(t *testing.T) {
+func TestForwarderDatadog_Call_Failure(t *testing.T) {
 	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusInternalServerError)
 	}))
@@ -55,23 +54,28 @@ func TestForwarderHttp_Call_Failure(t *testing.T) {
 	forwarder := &models.ForwarderV2{
 		Version: 2,
 		Config: models.ForwarderConfigV2{
-			Http: &models.ForwarderHttpV2{
-				Url:     testServer.URL,
-				Headers: map[string]string{},
+			Datadog: &models.ForwarderDatadogV2{
+				Url:    testServer.URL,
+				ApiKey: "apiKey",
 			},
 		},
 	}
 
-	if err := forwarder.Init(t.Context()); err != nil {
+	runtime, err := forwarders.NewRuntime(forwarder)
+	if err != nil {
+		t.Fatalf("failed to create runtime: %v", err)
+	}
+
+	if err := runtime.Init(t.Context()); err != nil {
 		t.Fatalf("failed to initialize forwarder: %v", err)
 	}
 
 	record := models.NewLogRecord(map[string]string{})
-	if err := forwarder.Call(t.Context(), record); err == nil {
+	if err := runtime.Call(t.Context(), record); err == nil {
 		t.Fatalf("expected error")
 	}
 
-	if err := forwarder.Close(t.Context()); err != nil {
+	if err := runtime.Close(t.Context()); err != nil {
 		t.Fatalf("failed to close forwarder: %v", err)
 	}
 }
